pyproject.toml
[tool.poetry]
name = "krishi-api"
version = "1.0.0"
description = "Python API for Krishi App"
authors = ["Himanshu Sagar <sagarhimanshu355@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
toml = "^0.10.2"
Flask = "^2.1.2"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
bson = "^0.5.10"
cmake = "^3.22.4"
psycopg2-binary = "^2.9.3"
psycopg2 = "^2.9.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
